+++ src/app/(auth)/signup/page.tsx
@@ -1,0 +1,12 @@
+import SignUp from '@/components/layout/auth/SignUp';
+import React from 'react';
+
+const ProtectedPage = () => {
+  return (
+    <>
+      <SignUp />
+    </>
+  );
+};
+
+export default ProtectedPage;
