+++ src/components/layout/customer/Dashboard.tsx
@@ -1,22 +1,9 @@
-'use client';
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 import ProtectedLayout from '../ProtectedLayout';
-import QuestionnairesModal from '@/components/QuestionnairesModal';
-import { trpc } from '@/utils/trpc';
 
 function CustomerDashboard() {
-  const [open, setOpen] = useState(false);
-  const { data: user, isLoading } = trpc.users.getUserById.useQuery();
-  useEffect(() => {
-    if (!isLoading && user) {
-      setOpen(user.questionnaires.length === 0);
-    }
-  }, [isLoading, user]);
-  console.log('user__', user);
-
   return (
     <ProtectedLayout>
-      <QuestionnairesModal isOpen={open} setIsOpen={setOpen} />
       Lorem ipsum dolor sit amet consectetur adipisicing elit. Perferendis
       accusamus recusandae officiis quisquam, rerum quasi vitae laudantium, rem
       provident mollitia fuga. Delectus corrupti eligendi nemo sunt officiis
