+++ src/server/db/interfaces/user.ts
@@ -1,5 +1,4 @@
 import { Document } from 'mongoose';
-
 interface IQuestionnaire {
   question: string; // E.g., 'Work and Education'
   answers: string[]; // Array of questions
